Introduce agent verification strategies
They allow verifying a VM prior to making it a Jenkins agent.
This way if a VM is not trusted we can delete it and create a new one.

Current strategies:
* Default - rely on VPN and SSH security
* AES - allow users to create a custom AES verification handshake

In addition, allow extensibility of the verification, meaning users can implement their own and plug them into the plugin.