(feat) Allow the user to add a prefix to the name of the different cloud agents (#59)
Currently the agents are named on the ID of the VM that stays behind them. With this feature, the user is allowed to specify a prefix in the agents config. The prefix then is added in front of the VM ID and the agent name looks like `prefix_vmid`. This gives the user a better way to distinguish between agents.