Add timeout to deleteVM operation's http client
The http client was instantiated with constructor that has hardcoded timeout
which was set to 0 (disabled) and this could block irrelevant node provisioning
operations for the duration of the API call to Orka. Adding a timeout should at
least ensure that another provisioning request will be processed after the
timeout is reached.