Take advantage of Orka API new token algorithm
Previously, the server was storing tons of tokens in a DB and we needed to revoke them to prevent DB pollution.
Now, there is a single token per user, so we no longer need to revoke them.

In addition, detect the server version and choose our approach base on that.