Fix overzealous auth token revocation
Whenever we perform a health check, we are recreating an `OrkaClient`.
If we're on Orka versions lower than 2.1.1, or it's the first time we
are performing the check after Jenkins restart, this forces the
reinitialization of the authentication token. This means that any active
tokens are revoked and therefore the users are logged out from all of
their other systems (such as CLI, web UI).

Because the health check doesn't actually need an authentication token,
just skip generating one for this case (and thereby skip revoking it
when the client is closed shortly).