Fix server timeout when deploying VMs (#126)
Orka server has a default timeout of 10 minutes. We can increase this timeout.
However, there isn't need for that in the plugin. We already have a request timeout for deployment, so we keep using it.
Make sure to increase the server timeout to a day, so it is always greater than the request timeout. Essentially we make sure than the server timeout never kicks in